+++ packages/turbo_bridge_mcp/lib/src/version_info.dart
@@ -1,4 +1,4 @@
-const String turboBridgeMcpVersion = '0.1.4';
+const String turboBridgeMcpVersion = '0.1.5';
 
 Map<String, dynamic> buildMcpCompatibilityInfo({
   required String bridgeVersion,